refactor: SEO
Moved SEO out of the `layouts`, since it technically isn't really part
of the *visible* layout. It is part of the page, though.

fixes: Issue with `Onepage` displaying "Contact" in the tab, because
that is the last layout to be displayed, thus the last layout to call
the SEO component.

note: This does not yet fix the issue of *Template pages correctly
providing SEO data. This is a fix for /pages/ only.