refactor!(markdown-pages): fileAbsolutePath routes
Removes the need for manual "path" in the markdown frontmatter,
the new implementation bases the generated URLs on the folder
structure.

This has also resulted in the move of all project images to be
colocated with the posts, making it more organized.

The duplicate names for folders / markdown files might be a bit
verbose, but the file names technically don't matter anymore. They
could all just as well be "index.md" or "post.md". The routeGenerator
ignores that part of the absolute path anyway.

The `TagTemplate` has now also become the `SkillsetTemplate`, as the
newly created pages will be under the "/skillset/" directory, making the
navigation still display as "active" on nested routes.

Next step would be to revise / clean up gatsby-node. The code in there
is ugly.