feat: added "uncooked" svgs from Affinity exports
Uncooked, as in "pre-Component". Irrelevant markup needs to be
manually changed / removed before it can become a `CodeSVG`, in
this case.

Wonder if there is an option to make Affinity export SVGs with a similar
naming format to just PNGs. So not adding "_<slice_number>" to the
filename. It doesn't when exporting PNGs.