feat(JSON-LD): initial Schema.org implementation
I am aware that this structured data does nothing for me with my data,
as per at least Google's search gallery:

https://developers.google.com/search/docs/guides/search-gallery

I was mostly just curious and thought it could be fun to try and play
with. The only thing that might work is the breadcrumb, but that alone
is not really worth the effort. Might decide at a later point to rip
the entire thing out again, who knows.