refactor(DataManager): reunited projectsData
There is no more ambiguity what "programmingData" is. This refactor
changes the data structure to be more predictable and in line with both
`experienceData` ánd `skillsetData`. The separation still exists.

Right now, the only thing that has changed is the fact that a "graphic
design project" now theoretically could be part of the "featured" top-3
projects on the homepage (and therefore also display an image). If I
decide to ditch featured images / videos in graphic design data I should
just remove them from the MDX files.